Refactor navigation and dagger 2, many other changes, debugging funct…

+++ muciplayer-data/src/main/java/com/andreich/data/datasource/remote/RemoteDataSource.kt
@@ -8,7 +8,7 @@ interface RemoteDataSource {
 
     suspend fun getChartTracks(): TracksDto
 
-    suspend fun getTrack(id: Int): TrackDetailDto
+    suspend fun getTrack(id: Long): TrackDetailDto
 
     suspend fun searchTrack(query: String): SearchResultDto
 }
